--- internal/x/urlx/encoded_slash_test.go
@@ -1,59 +1,0 @@
-// Copyright 2026 Dimitrij Drus <dadrus@gmx.de>
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-//
-// SPDX-License-Identifier: Apache-2.0
-
-package urlx
-
-import (
-	"testing"
-
-	"github.com/stretchr/testify/assert"
-)
-
-func TestContainsEncodedSlash(t *testing.T) {
-	t.Parallel()
-
-	for uc, tc := range map[string]struct {
-		path     string
-		expected bool
-	}{
-		"empty": {},
-		"without escapes": {
-			path: "/api/v1/resource",
-		},
-		"uppercase sequence": {
-			path:     "/api%2Fv1/resource",
-			expected: true,
-		},
-		"lowercase sequence": {
-			path:     "/api%2fv1/resource",
-			expected: true,
-		},
-		"mixed in long path": {
-			path:     "/foo/bar/baz%2Fqux/quux",
-			expected: true,
-		},
-		"not slash escape": {
-			path: "/api%2Ev1/resource",
-		},
-		"incomplete escape": {
-			path: "/api%2",
-		},
-	} {
-		t.Run(uc, func(t *testing.T) {
-			assert.Equal(t, tc.expected, ContainsEncodedSlash(tc.path))
-		})
-	}
-}
